+++ packages/decorator/src/common/constant.ts
@@ -1,3 +1,6 @@
+// got all value with no property name
+export const ALL_VALUE= 'common:all_value_key';
+
 // common
 export const PRIORITY_KEY = 'common:priority';
 export const SCHEDULE_KEY = 'common:schedule';