+++ packages/web/test/fixtures/feature/base-app-middleware/src/config/config.default.ts
@@ -1,0 +1,9 @@
+export const keys = 'key';
+
+export const hello = {
+  a: 1,
+  b: 2,
+  d: [1, 2, 3],
+};
+
+export const middleware = ['globalMiddleware2'];