+++ packages/web-koa/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.4](https://github.com/midwayjs/midway/compare/v2.2.3...v2.2.4) (2020-09-15)
+
+
+### Bug Fixes
+
+* support midway global middleware use id ([8dc9ae3](https://github.com/midwayjs/midway/commit/8dc9ae33acd559d74f144a75f08fc039037fa45b))
+
+
+
+
+
+## [2.2.3](https://github.com/midwayjs/midway/compare/v2.2.2...v2.2.3) (2020-09-14)
+
+**Note:** Version bump only for package @midwayjs/koa
+
+
+
+
+
+## [2.2.2](https://github.com/midwayjs/midway/compare/v2.2.1...v2.2.2) (2020-09-14)
+
+**Note:** Version bump only for package @midwayjs/koa
+
+
+
+
+
+## [2.2.1](https://github.com/midwayjs/midway/compare/v2.2.0...v2.2.1) (2020-09-14)
+
+**Note:** Version bump only for package @midwayjs/koa
+
+
+
+
+
 # [2.2.0](https://github.com/midwayjs/midway/compare/v2.1.4...v2.2.0) (2020-09-13)